+++ Components/IconTypes/IconType_item/item.lua
@@ -130,7 +130,7 @@ local function ItemCooldown_OnUpdate(icon, time)
 		start, duration, enable = item:GetCooldown()
 
 		if duration then
-			if enable == 0 then
+			if enable == 0 or enable == false then
 				-- Enable will be 0 for things like a potion that was used in combat 
 				-- and the cooldown hasn't yet started counting down.
 				start, duration = 0, 0
@@ -146,7 +146,7 @@ local function ItemCooldown_OnUpdate(icon, time)
 				equipped = false
 			end
 			
-			if equipped and inrange and enable == 1 and (duration == 0 or OnGCD(duration)) then
+			if equipped and inrange and (enable == 1 or enable == true) and (duration == 0 or OnGCD(duration)) then
 				-- This item is usable. Set the attributes and then stop.
 
 				icon:SetInfo("state; texture; start, duration; stack, stackText; spell",